requirements.txt
keyboard==0.13.5
mss==9.0.1
numpy==1.26.2; python_version < "3.13"
numpy>=2.2,<3; python_version >= "3.13"
openai==1.6.1
Pillow==10.1.0; python_version < "3.13"
Pillow>=11.0,<13; python_version >= "3.13"
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pynput==1.7.6
PyQt5==5.15.10; python_version < "3.13"
PyQt5>=5.15.11,<6; python_version >= "3.13"
PyQt5_sip==12.13.0; python_version < "3.13"
PyQt5_sip>=12.15,<13; python_version >= "3.13"
pytesseract==0.3.10
python-dotenv==1.0.0
scikit_learn==1.3.2; python_version < "3.13"
scikit_learn>=1.6,<2; python_version >= "3.13"
termcolor==2.4.0
opencv-python>=4.10
